[feat] #227 Remove CardBackground View
